working_demo/requirements.txt
beautifulsoup4==4.7.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Dropzone==1.5.3
Flask-Mail==0.9.1
Flask-WTF==0.14.2
lxml==4.3.2
mysql-connector==2.2.9
numpy==1.16.2
opencv-python==4.1.0.25
pytesseract==0.2.6
sshtunnel==0.1.4